pyproject.toml
[tool.poetry]
name = "gork"
version = "0.1.0"
description = "Make images accessible"
authors = ["Gökmen Görgen <gkmngrgn@gmail.com>"]
license = "MIT"

[tool.poetry.scripts]
gork = "gork.main:run_cli"

[tool.poetry.dependencies]
python = "^3.8"
pillow = "7.0.0"
numpy = "1.18.1"

[tool.poetry.dev-dependencies]
mypy = "^0.761"
black = "^19.10b0"
isort = "^4.3.21"
pytest = "5.4.1"
ipdb = "^0.13.2"
coverage = {version = "5.1", extras = ["toml"]}

[tool.black]
line-length = 119
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | venv
    | frontend
  )/
  | settings.py
)
'''

[tool.coverage.run]
source = ["./gork"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
